Refactor runtime into dedicated modules
