+++ crates/compose-core/src/frame_clock.rs
@@ -1,0 +1,75 @@
+use crate::runtime::RuntimeHandle;
+use crate::FrameCallbackId;
+
+#[derive(Clone)]
+pub struct FrameClock {
+    runtime: RuntimeHandle,
+}
+
+impl FrameClock {
+    pub fn new(runtime: RuntimeHandle) -> Self {
+        Self { runtime }
+    }
+
+    pub fn runtime_handle(&self) -> RuntimeHandle {
+        self.runtime.clone()
+    }
+
+    pub fn with_frame_nanos(
+        &self,
+        callback: impl FnOnce(u64) + 'static,
+    ) -> FrameCallbackRegistration {
+        let mut callback_opt = Some(callback);
+        let runtime = self.runtime.clone();
+        match runtime.register_frame_callback(move |time| {
+            if let Some(callback) = callback_opt.take() {
+                callback(time);
+            }
+        }) {
+            Some(id) => FrameCallbackRegistration::new(runtime, id),
+            None => FrameCallbackRegistration::inactive(runtime),
+        }
+    }
+
+    pub fn with_frame_millis(
+        &self,
+        callback: impl FnOnce(u64) + 'static,
+    ) -> FrameCallbackRegistration {
+        self.with_frame_nanos(move |nanos| {
+            let millis = nanos / 1_000_000;
+            callback(millis);
+        })
+    }
+}
+
+pub struct FrameCallbackRegistration {
+    runtime: RuntimeHandle,
+    id: Option<FrameCallbackId>,
+}
+
+impl FrameCallbackRegistration {
+    fn new(runtime: RuntimeHandle, id: FrameCallbackId) -> Self {
+        Self {
+            runtime,
+            id: Some(id),
+        }
+    }
+
+    fn inactive(runtime: RuntimeHandle) -> Self {
+        Self { runtime, id: None }
+    }
+
+    pub fn cancel(mut self) {
+        if let Some(id) = self.id.take() {
+            self.runtime.cancel_frame_callback(id);
+        }
+    }
+}
+
+impl Drop for FrameCallbackRegistration {
+    fn drop(&mut self) {
+        if let Some(id) = self.id.take() {
+            self.runtime.cancel_frame_callback(id);
+        }
+    }
+}
